+++ browser/static/ably.d.ts
@@ -332,7 +332,7 @@ declare namespace Types {
     syncComplete: boolean;
     get: (Params: realtimePresenceGetCallback | RealtimePresenceParams, callback?: realtimePresenceGetCallback) => void;
     history: (ParamsOrCallback: RealtimePresenceHistoryParams | paginatedResultCallback<PresenceMessage>, callback?: paginatedResultCallback<PresenceMessage>) => void;
-    subscribe: (presenceOrCallback: PresenceAction | messageCallback<PresenceMessage>, listener?: messageCallback<PresenceMessage>) => void;
+    subscribe: (presenceOrCallback: PresenceAction | messageCallback<PresenceMessage> | Array<PresenceAction>, listener?: messageCallback<PresenceMessage>) => void;
     unsubscribe: (presence?: PresenceAction, listener?: messageCallback<PresenceMessage>) => void;
     enter: (data?: errorCallback | any, callback?: errorCallback) => void;
     update: (data?: errorCallback | any, callback?: errorCallback) => void;
@@ -357,7 +357,7 @@ declare namespace Types {
     attach: (callback?: standardCallback) => void;
     detach: (callback?: standardCallback) => void;
     history: (paramsOrCallback?: RealtimePresenceHistoryParams | paginatedResultCallback<Message>, callback?: paginatedResultCallback<Message>) => void;
-    subscribe: (eventOrCallback: messageCallback<Message> | string, listener?: messageCallback<Message>) => void;
+    subscribe: (eventOrCallback: messageCallback<Message> | string | Array<string>, listener?: messageCallback<Message>) => void;
     unsubscribe: (eventOrCallback?: messageCallback<Message> | string, listener?: messageCallback<Message>) => void;
     publish: (messagesOrName: any, messageDataOrCallback?: errorCallback | any, callback?: errorCallback) => void;
     setOptions: (options: any, callback?: errorCallback) => void;